% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateHierarchicalPrediction.R
\name{annotateHierarchicalPrediction}
\alias{annotateHierarchicalPrediction}
\title{Annotate Hierarchical Prediction}
\usage{
annotateHierarchicalPrediction(x, ...)
}
\description{
Annotate Hierarchical Prediction
}
