% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureContrib.R
\name{featureContrib}
\alias{featureContrib}
\title{Feature Contribution}
\usage{
featureContrib(tidy.RF, testX)
}
\arguments{
\item{tidy.RF}{A tidy random forest. The random forest to make predictions
with.}

\item{testX}{A data frame. Test set features.}
}
\value{
A list. The structure and content depends the type of the forest.
  \itemize{
    \item Regression For each observation, a P-by-1 matrix, where P is the
      number of features in \code{testX}. Each row of the matrix stands for
      the contribution of that feature to the prediction of the response.
    \item Classification For each observation, a P-by-K matrix, where P is
      the number of features in \code{testX}, and K is the number of
      response classes. Each row of the matrix stands for the contribution
      of that feature to the prediction of each response class.
  }
}
\description{
Contribution of each feature to the prediction of test data.
}
\details{
Recall that each node in a decision tree has a prediction associated with
it. For regression trees, it's the average response in that node, whereas
in classification trees, it's the frequency of each response class, or the
most frequent response class in that node.

For an observation and a tree in the forest, the contribution of each
feature is the sum of differences between the predictions of nodes which
split on it and those of their children, i.e. the sum of changes in node
prediction caused by spliting on it.

For an observation and a forest, the contribution of each feature is simply
the average contribution across all trees. This is because the prediction of
a forest is the average of the predictions of its trees.

Together with \code{trainsetBias}, it can perform a decomposition of the
prediction for each observation by feature importance:

prediction(tree, X) = trainsetBias(tree) + featureContrib_1(tree, X) +
                featureContrib_2(tree, X) + ... + featureContrib_p(tree, X)
}
\references{
\url{http://blog.datadive.net/interpreting-random-forests/}

\url{http://blog.datadive.net/random-forest-interpretation-with-scikit-learn/}
}
\seealso{
\code{\link{trainsetBias}}
}
