% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureContrib.R
\name{trainsetBias}
\alias{trainsetBias}
\title{Trainset Bias}
\usage{
trainsetBias(tidy.RF)
}
\arguments{
\item{tidy.RF}{A tidy random forest. The random forest to extract train
set bias from.}
}
\value{
A list. The structure and content depends the type of the forest.
  \itemize{
    \item Regression: A 1-by-1 matrix. The trainset bias for the prediction
      of the response.
    \item Classification: A 1-by-K matrix, where K is the number of
      response classes. Each column of the matrix stands for the trainset
      bias for the prediction of each response class.
  }
}
\description{
For a tree in the forest, trainset bias is the prediction of its root node,
or the unconditional prediction of the tree, or the average response of its
in-bag observations.
}
\details{
For a forest, the trainset bias is simply the average trainset bias across
all trees. This is because the prediction of a forest is the average of the
predictions of its trees.

Together with \code{trainsetBias}, it can perform a decomposition of the
prediction for each observation by feature importance:

prediction(tree, X) = trainsetBias(tree) + featureContrib_1(tree, X) +
                featureContrib_2(tree, X) + ... + featureContrib_p(tree, X)
}
\references{
\url{http://blog.datadive.net/interpreting-random-forests/}

\url{http://blog.datadive.net/random-forest-interpretation-with-scikit-learn/}
}
\seealso{
\code{\link{featureContrib}}
}
